+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/artifacts/EtherealChains.java
@@ -282,7 +282,7 @@ public class EtherealChains extends Artifact {
 	@Override
 	public void charge(Hero target, float amount) {
 		if (cursed || target.buff(MagicImmune.class) != null) return;
-		long chargeTarget = (long) ((5+(level()*2))*getRarityMultiplier());
+		long chargeTarget = ((5+(level()*2))*getRarityMultiplier());
 		if (charge < chargeTarget*2){
 			partialCharge += 0.5f*amount;
 			partialCharge *= getRarityMultiplier();
