+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/wands/WandOfFrost.java
@@ -51,11 +51,11 @@ public class WandOfFrost extends DamageWand {
 	}
 
 	public long min(long lvl){
-		return (long) ((2+lvl)*getRarityMultiplier());
+		return ((2+lvl)*getRarityMultiplier());
 	}
 
 	public long max(long lvl){
-		return (long) ((8+5*lvl)*getRarityMultiplier());
+		return ((8+5*lvl)*getRarityMultiplier());
 	}
 
 	@Override
