+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/wands/WandOfLightning.java
@@ -61,11 +61,11 @@ public class WandOfLightning extends DamageWand {
 	private ArrayList<Lightning.Arc> arcs = new ArrayList<>();
 
 	public long min(long lvl){
-		return (long) ((5+lvl)*getRarityMultiplier());
+		return ((5+lvl)*getRarityMultiplier());
 	}
 
 	public long max(long lvl){
-		return (long) ((10+5*lvl)*getRarityMultiplier());
+		return ((10+5*lvl)*getRarityMultiplier());
 	}
 	
 	@Override
