+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/wands/WandOfPrismaticLight.java
@@ -61,11 +61,11 @@ public class WandOfPrismaticLight extends DamageWand {
 	}
 
 	public long min(long lvl){
-		return (long) ((1+lvl)*getRarityMultiplier());
+		return ((1+lvl)*getRarityMultiplier());
 	}
 
 	public long max(long lvl){
-		return (long) ((5+3*lvl)*getRarityMultiplier());
+		return ((5+3*lvl)*getRarityMultiplier());
 	}
 
 	@Override
